# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a service for ePharma project",
        "title": "ePharma API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "10.225.78.198:8080",
    "basePath": "/",
    "paths": {
...
        "/product": {
            "get": {
                "description": "get products",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Get all products",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_tadilbek11kz_ePharma-backend_pkg_product.Product"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "description": "create product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Create a product",
                "parameters": [
                    {
                        "description": "body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_tadilbek11kz_ePharma-backend_pkg_product.CreateProductRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_tadilbek11kz_ePharma-backend_pkg_product.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/product/{id}": {
            "get": {
                "description": "get product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Get a product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_tadilbek11kz_ePharma-backend_pkg_product.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "description": "update product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Update a product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_tadilbek11kz_ePharma-backend_pkg_product.UpdateProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_tadilbek11kz_ePharma-backend_pkg_product.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "description": "delete product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Delete a product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/product/{id}/inventory": {
            "get": {
                "description": "get product inventory",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "Get product inventory",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_tadilbek11kz_ePharma-backend_pkg_pharmacy.Pharmacy"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_tadilbek11kz_ePharma-backend_pkg_pharmacy.Pharmacy": {
            "type": "object",
            "required": [
                "location",
                "name"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "descritpion": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "github_com_tadilbek11kz_ePharma-backend_pkg_product.CreateProductRequest": {
            "type": "object",
            "required": [
                "brand_name",
                "dispence_mode",
                "dosage",
                "generic_name",
                "image",
                "insurance_plan",
                "manufacturer_name",
                "package_size",
                "strength"
            ],
            "properties": {
                "brand_name": {
                    "type": "string"
                },
                "dispence_mode": {
                    "type": "string"
                },
                "dosage": {
                    "type": "string"
                },
                "generic_name": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "insurance_plan": {
                    "type": "string"
                },
                "manufacturer_name": {
                    "type": "string"
                },
                "package_size": {
                    "type": "string"
                },
                "strength": {
                    "type": "string"
                }
            }
        },
        "github_com_tadilbek11kz_ePharma-backend_pkg_product.Product": {
            "type": "object",
            "required": [
                "brand_name",
                "dispence_mode",
                "dosage",
                "generic_name",
                "image",
                "insurance_plan",
                "manufacturer_name",
                "package_size",
                "strength"
            ],
            "properties": {
                "brand_name": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "dispence_mode": {
                    "type": "string"
                },
                "dosage": {
                    "type": "string"
                },
                "generic_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "insurance_plan": {
                    "type": "string"
                },
                "manufacturer_name": {
                    "type": "string"
                },
                "package_size": {
                    "type": "string"
                },
                "strength": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_tadilbek11kz_ePharma-backend_pkg_product.UpdateProductRequest": {
            "type": "object",
            "properties": {
                "brand_name": {
                    "type": "string"
                },
                "dispence_mode": {
                    "type": "string"
                },
                "dosage": {
                    "type": "string"
                },
                "generic_name": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "insurance_plan": {
                    "type": "string"
                },
                "manufacturer_name": {
                    "type": "string"
                },
                "package_size": {
                    "type": "string"
                },
                "strength": {
                    "type": "string"
                }
            }
        },
...
    }
}